+++ changelog.md
@@ -4,6 +4,7 @@
 
 1. Added TruckersMP user info in applications
 2. Added export member list (csv)
+3. Added login screen tips
 
 ## v3.4.1